move get_lut_cost() to util.py
