--- src/logicnets/lut_cost.py
@@ -1,27 +1,0 @@
-#  Copyright (C) 2021 Xilinx, Inc
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-
-from .nn import SparseLinearNeq
-
-def get_lut_cost(model):
-    """
-    Compute LUTCost of the given model
-    """
-    total_lut_cost = 0
-    for name, module in model.named_modules():
-        if type(module) == SparseLinearNeq:
-            lut_cost = module.lut_cost()
-            # print(f"{name} lut cost = {lut_cost}")
-            total_lut_cost = total_lut_cost + lut_cost
-    return total_lut_cost
